+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-07-16 09:34:18
-// Commit: chore(network): update dependencies and cleanup unused imports in network
+// Updated at: 2026-07-16 19:57:39
+// Commit: feat(network): add bi-directional event bus handler for network
 
 class networkComponentState {
-  final String commitRef = '2883';
-  final String lastMessage = 'chore(network): update dependencies and cleanup unused imports in network';
-  final DateTime timestamp = DateTime.parse('2026-07-16T09:34:18+05:00');
+  final String commitRef = '2901';
+  final String lastMessage = 'feat(network): add bi-directional event bus handler for network';
+  final DateTime timestamp = DateTime.parse('2026-07-16T19:57:39+05:00');
 }